chore: fix casing part 2/2
Git is case sensitive, MacOS is not. A couple of commits ago, I renamed
Topic.ts to topic.ts. But doing it with the mv command resulted in the
change not registering in git. So the code compiled on my machine, but
broke as soon as I checked out the repo in another location (i.e.
travis).

Maybe I could have done it smoother using git mv.