feat(deploy): add topic flag
Adds the option to specify a single topic to deploy. This works by
writing a temporary root config to root.config.local.yml, containing
only the top level entry for the selected topic. The ansible playbook is
configured to accept overrides contained in this file. It will first
source the original root config, storing all topic configurations in the
'topics' variable. This is overwritten in root.config.local.yml.

I should be careful not to enable the ansible option that alters how
maps are merged... that would result in unexpected results, I think.